requirements.txt
aniso8601==9.0.1
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.1
click==8.0.1
Flask==2.0.1
flask-marshmallow==0.14.0
lxml==4.6.2
MarkupSafe==2.0.1
marshmallow==3.12.2
mccabe==0.6.1
psycopg2-binary==2.9
pycodestyle==2.7.0
python-dotenv~=0.18.0
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
Werkzeug==2.0.1
pyjwt
pytest

